requirements.txt
streamlit==1.29.0
openai==1.3.8
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.1.0